loader: Move merge helper to internal pkg
Make the private helper exposed internally as a utility function to
help merge data trees.
